tests/wpt/web-platform-tests/css/css-multicol/parsing/column-fill-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-fill with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-fill">
<meta name="assert" content="column-fill supports only the grammar 'auto | balance | balance-all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("column-fill", "none");
test_invalid_value("column-fill", "auto balance");
</script>
</body>
</html>
